Create school-data-parser.
Script creates school objects.  Iterates through objects to find schools with the same address -- places these in CSV.  Then iterates through csv to create another csv with schools with charter schools and a hash which groups schools by address.  End result of script is a hash with address keys and list of schools ad values if the address has a charter.  There are 99 co-located charter schools in the city per this script.